mblog_api/requirements.txt
aiohttp==3.6.2
aiohttp-session==2.9.0
aiohttp-security==0.4.0
aiopg==1.0.0
click==7.0
cryptography==2.8
python-rapidjson==0.9.1
sqlalchemy==1.3.11
uvloop==0.14.0